deploy v1

+++ src/main/java/com/back_end/english_app/repository/ForumPostViewRepository.java
@@ -11,6 +11,6 @@ public interface ForumPostViewRepository extends JpaRepository<ForumPostViewEnti
 
     boolean existsByPostIdAndIpAddressAndViewedAtAfter(Long postId, String ipAddress, java.time.LocalDateTime after);
 
-    long countByPostId(Long postId);
+    Integer countByPostId(Long postId);
 }
